+++ src/components/ExtendedMultiselectInput.vue
@@ -42,7 +42,7 @@
       :style="multipleBlocksMargin"
       :disabled="disabled"
       :loading="loading"
-      :icon-filter="iconFilter"
+      :icon-filter="loaderIconFilter"
       :placeholder-block-shown="!selectedOptions.length"
       :show-deselect-icon-loader="showDeselectIconLoader"
       :toggle-multiple-blocks-limit="toggleMultipleBlocksLimit"
@@ -102,7 +102,7 @@
     <extended-multiselect-multiple
       :disabled="disabled"
       :loading="loading"
-      :icon-filter="iconFilter"
+      :icon-filter="loaderIconFilter"
       :show-deselect-icon-loader="showDeselectIconLoader"
       :toggle-multiple-blocks-limit="toggleMultipleBlocksLimit"
       :empty-objects-placeholder="emptyObjectsPlaceholder"
@@ -333,10 +333,10 @@ export default Vue.extend({
     },
 
     /**
-     * Defines a svg-filter for icons
-     * @property {string} iconFilter
+     * Defines a svg-filter for loader icons
+     * @property {string} loaderIconFilter
      */
-    iconFilter: {
+    loaderIconFilter: {
       type: String,
       required: true,
     },